+++ src/main/regression.d
@@ -24,7 +24,7 @@ void main(string[] args){
   dvector trait = [1,1,1,2000,2000,2000];
   dvector weight = [1,1,1,1,1,1];
   ivector nullmodellayout = [1,1];  //The D[][1] is dropped from the model to test its predictive value 
-  for(uint i=0; i<designmatrix.length; i++){
+  for(size_t i = 0; i < designmatrix.length; i++){
     writefln("[%f] = %f",trait[i],designmatrix[i]);
   }
   writefln("LOD = %f",multipleregression(designmatrix,trait,weight,nullmodellayout,0));
